Many changes: single frame gfx, match time, tweaks, random soundtrack.
The character is now render as a "frame graphic" that is specified in a
text file with a name, and some parameters.  Next set is to get frames
from a manager and render a different frame for different
states/actions.

At the end of the the match the match time in seconds is printed to the
console.

There were some gameplay tweaks (up tilt has a faster recovery time).

On startup a random track from a list of hardcoded tracks is selected.